feat: extract utility
